+++ test/Sentry.AspNetCore.Grpc.Tests/SentryCoreDependentCollection.cs
@@ -1,4 +1,3 @@
-using Sentry.AspNetCore.Tests;
 using Xunit;
 
 namespace Sentry.AspNetCore.Grpc.Tests