+++ test/Sentry.Extensions.Logging.Tests/SentryLoggerFactoryExtensionsTests.cs
@@ -2,8 +2,6 @@ using System;
 using Microsoft.Extensions.Logging;
 using NSubstitute;
 using Sentry.Extensibility;
-using Sentry.Internal;
-using Sentry.Protocol;
 using Xunit;
 
 namespace Sentry.Extensions.Logging.Tests