+++ test/Sentry.Tests/SentryClientTests.cs
@@ -5,10 +5,8 @@ using System.Threading.Tasks;
 using NSubstitute;
 using Sentry.Extensibility;
 using Sentry.Internal;
-using Sentry.Protocol;
 using Sentry.Protocol.Envelopes;
 using VerifyXunit;
-using VerifyTests;
 using Xunit;
 
 namespace Sentry.Tests